quell warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17239 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/net/multicast.cxx
@@ -72,7 +72,7 @@ int			openBroadcast(int port, const char* service,
   }
 
   /* open socket */
-  fd = socket(AF_INET, SOCK_DGRAM, 0);
+  fd = (int)socket(AF_INET, SOCK_DGRAM, 0);
   if (fd < 0) {
     nerror("openBroadcast: socket");
     return -1;
